+++ server/go/internal/win/win32/msi/msi.go
@@ -1,0 +1,11 @@
+//go:build windows && amd64
+
+package msi
+
+import (
+	"golang.org/x/sys/windows"
+)
+
+var (
+	modmsi = windows.NewLazySystemDLL("msi.dll")
+)